First cut at multilayer optimization via heights
In the new method, shown in metasurface_optimization.ipynb and nearfield_optimization.ipynb, float heights of each pixel are directly optimized, leading to a scalable method for many device layers. The old method, with one layer optimized at a time, is for now preserved in the *_singlelayer.ipynb and *_multilayer.ipynb notebooks.